testiranje, stranica za menjanje lozinke prilikom prvog
prijavljivanja na sistem, rad na prikazu za radnike

+++ src/main/java/com/example/model/TableRestaurant.java
@@ -10,7 +10,7 @@ import javax.persistence.ManyToOne;
 @Entity
 public class TableRestaurant {
 	@Id
-	private int tableNumber;
+	private long tableNumber;
 	
 	@Column(nullable = false)
 	private String segment;
@@ -53,11 +53,11 @@ public class TableRestaurant {
 		this.isAvailable = isAvailable;
 	}
 
-	public int getTableNumber() {
+	public long getTableNumber() {
 		return tableNumber;
 	}
 
-	public void setTableNumber(int tableNumber) {
+	public void setTableNumber(long tableNumber) {
 		this.tableNumber = tableNumber;
 	}
