+++ src/main/java/com/example/repository/TableRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.repository.Repository;
 
 import com.example.model.TableRestaurant;
 
-public interface TableRepository extends Repository<TableRestaurant, Integer> {
+public interface TableRepository extends Repository<TableRestaurant, Long> {
 	
 	public TableRestaurant save(TableRestaurant entity);
 	
-	public TableRestaurant findOne(int id);
+	public TableRestaurant findOne(long id);
 	
 	public List<TableRestaurant> findByRestaurant(long restaurant);
 }
